backend/pyproject.toml
[project]
name = "engin33ring-thesis"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "boto3>=1.42.59",
    "boto3-stubs>=1.42.59",
    "botocore>=1.42.59",
    "fastapi>=0.128.0",
    "mypy>=1.19.1",
    "mypy-boto3>=1.42.3",
    "pydantic-settings[yaml]>=2.12.0",
    "pymongo>=4.16.0",
    "pytest-asyncio>=1.3.0",
    "pytest-mock>=3.15.1",
    "python-multipart>=0.0.22",
    "ruff>=0.14.13",
    "uvicorn>=0.40.0",
    "prometheus-fastapi-instrumentator>=7.0.0",
    "python-json-logger>=3.2.1",
    "opentelemetry-api>=1.30.0",
    "opentelemetry-sdk>=1.30.0",
    "opentelemetry-exporter-otlp>=1.30.0",
    "opentelemetry-instrumentation-logging>=0.51b0",
    "opentelemetry-instrumentation-fastapi>=0.51b0",
    "ffmpeg-python>=0.2.0",
    "pyjwt>=2.12.1",
    "pwdlib[argon2]>=0.3.0",
    "pytest>=9.0.2",
    "pillow>=12.2.0",
]

[tool.mypy]
plugins = ['pydantic.mypy']

[[tool.mypy.overrides]]
module = ["backend.*", "ffmpeg.*"]
ignore_missing_imports = true

[tool.ruff]
line-length = 88
indent-width = 4

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
